slack connector: implemented signing secret and updated docs (#7186)
* implemented signing secret and updated docs

* added changelog

* Update docs/docs/connectors/slack.mdx

* fixed missing file extension

* fixed tests

* Apply suggestions from code review


* added slack channel tests

* Update slack.py

* Apply suggestions from code review


* implemented review comments
