+++ changelog/7186.bugfix.md
@@ -1,0 +1,1 @@
+Properly validate incoming webhook requests for the Slack connector to be authentic.
