+++ rasa/shared/constants.py
@@ -13,6 +13,7 @@ DOCS_URL_POLICIES = DOCS_BASE_URL + "/policies"
 DOCS_URL_TEST_STORIES = DOCS_BASE_URL + "/testing-your-assistant"
 DOCS_URL_ACTIONS = DOCS_BASE_URL + "/actions"
 DOCS_URL_CONNECTORS = DOCS_BASE_URL + "/connectors/"
+DOCS_URL_CONNECTORS_SLACK = DOCS_URL_CONNECTORS + "/slack"
 DOCS_URL_EVENT_BROKERS = DOCS_BASE_URL + "/event-brokers"
 DOCS_URL_PIKA_EVENT_BROKER = DOCS_URL_EVENT_BROKERS + "#pika-event-broker"
 DOCS_URL_TRACKER_STORES = DOCS_BASE_URL + "/tracker-stores"
